[BREAKING] Convert LAMCommandExecutor to abstract class
