+++ src/cc/chaoticweg/mc/LeaveAMessage/LAMPlugin.java
@@ -1,6 +1,6 @@
 package cc.chaoticweg.mc.LeaveAMessage;
 
-import cc.chaoticweg.mc.LeaveAMessage.command.LAMCommandExecutor;
+import cc.chaoticweg.mc.LeaveAMessage.command.LAMMsgCommandExecutor;
 import cc.chaoticweg.mc.LeaveAMessage.database.LAMDatabaseHandler;
 import cc.chaoticweg.mc.LeaveAMessage.database.LAMMessage;
 import cc.chaoticweg.mc.LeaveAMessage.event.LAMLoginEventListener;
@@ -33,7 +33,7 @@ public class LAMPlugin extends JavaPlugin {
         databaseHandler.setup(); // don't forget this, lol
 
         // set command and event handlers
-        this.getCommand("msg").setExecutor(LAMCommandExecutor.getInstance(this));
+        this.getCommand("msg").setExecutor(new LAMMsgCommandExecutor(this));
         this.getServer().getPluginManager().registerEvents(LAMLoginEventListener.getInstance(this), this);
 
         // done.
